Change jump integrator away from innovation form
Using the innovation results in all kinds of numerical problems,
such as having divisions by zero, that can be avoided. Now we have
two update expressions, one for no jump and one for jump, and the
expression is randomly chosen according to the jump rate at each
timestep, avoiding a bunch of numerical problems.