Cast traces as real vecs to avoid ComplexWarnings
Tracing a hermitian operator together with a hermitian operator
gives real results, so the dual vector representing such a trace
functional is real. In general dual vectors can be complex, so to
keep everything explicitly real in our integrators we have to
explicitly take the real part of the dual vector.