Add uncond int method with vec init cond
Allows one to specify the initial condition of a hierarchy
unconditional master equation in vectorized form. Useful if you
need to express a non-separable initial condition, such as when
using the quantum regression theorem, or if you already know the
vectorized form and don't want to waste time going back and forth
between density-matrix representations.