Add comments to sparse basis and improve dualize
Before, dualize was doing a silly effective contraction with the inverse
metric by using vectorize (which gets components using the dual basis),
requiring a subsequent contraction with the metric. Now dualize avoids
this contraction by explicitly cancelling the contraction of the metric
and inverse metric and simply getting components by contracting with the
vector basis.